Add files via upload
Topics that I have already started, plan to do, or feel well within my comfort zone to do in the future are all marked in red. There are a few other topics on there I can tackle with a little extra help or by more or less stitching them together with Max Cookbook examples, though I may not be able to write the script from scratch.