enforcing consistency on topics pages
enforcing consistency on topics pages.
removed last edited line on one or two of these and made sure all of them had summary section and ordered list. Intro to html and javascript is the only exception right now because it has "expectations and pre requisites" section on the topic page